requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
crewai==0.1.0
slack-bolt==1.18.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
pydantic==2.5.2
email-validator==2.1.0.post1
langchain-google-genai==0.0.5
google-generativeai==0.3.1
python-multipart==0.0.6 